// packages/web-browser/constants.d.ts
export declare const NEW_TAB_URL = "about:newtab";
export declare const ABOUT_SETTINGS_URL = "about:settings";
export declare const ICONS: {
    CLOSE: import("react/jsx-runtime").JSX.Element;
    SPARKLES: import("react/jsx-runtime").JSX.Element;
    WINDOW_MINIMIZE: import("react/jsx-runtime").JSX.Element;
    WINDOW_MAXIMIZE: import("react/jsx-runtime").JSX.Element;
    WINDOW_CLOSE: import("react/jsx-runtime").JSX.Element;
    ADD: import("react/jsx-runtime").JSX.Element;
    CHEVRON_LEFT: import("react/jsx-runtime").JSX.Element;
    CHEVRON_RIGHT: import("react/jsx-runtime").JSX.Element;
    RELOAD: import("react/jsx-runtime").JSX.Element;
    HOME: import("react/jsx-runtime").JSX.Element;
    MENU: import("react/jsx-runtime").JSX.Element;
    BOOKMARK: import("react/jsx-runtime").JSX.Element;
    BOOKMARK_FILLED: import("react/jsx-runtime").JSX.Element;
    LOCK: import("react/jsx-runtime").JSX.Element;
    INFO: import("react/jsx-runtime").JSX.Element;
    VERTICAL_TABS: import("react/jsx-runtime").JSX.Element;
};
//# sourceMappingURL=constants.d.ts.map